% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_priogrid.R
\name{pg_col}
\alias{pg_col}
\title{The column (int) of the PRIOGrid structure corresponding to a longitude (float)}
\usage{
pg_col(lon)
}
\arguments{
\item{float}{A longitude provided as a float (eg. -1.25). Longitude must be expressed as WGS84, i.e. [-180..+180]}
}
\value{
integer The corresponding PRIOGrid column id
}
\description{
The column (int) of the PRIOGrid structure corresponding to a longitude (float)
}
\examples{
pg_col(27.1)
pg_col(-24.5)
}
