% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_priogrid.R
\name{pg_priogrid}
\alias{pg_priogrid}
\title{The PRIOGrid GID (int) of the corresponding to a pair of coordinates (latitude (float) and longitude (float)).
Compatible with PRIOGrid v.1.x and v.2.x}
\usage{
pg_priogrid(lat, lon)
}
\arguments{
\item{float}{A latitude provided as a float (eg. -1.25). Latitude must be expressed as WGS84, i.e. [-90..+90]}

\item{float}{A longitude provided as a float (eg. -1.25). Longitude must be expressed as WGS84, i.e. [-180..+180]}
}
\value{
integer The corresponding PRIOGrid GID
}
\description{
The PRIOGrid GID (int) of the corresponding to a pair of coordinates (latitude (float) and longitude (float)).
Compatible with PRIOGrid v.1.x and v.2.x
}
\examples{
pg_priogrid(27.1,11.2)
pg_priogrid(-24.5.8.21)
}
