% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_priogrid.R
\name{pg_row}
\alias{pg_row}
\title{The row (int) of the PRIOGrid structure corresponding to a latitude (float)}
\usage{
pg_row(lat)
}
\arguments{
\item{float}{A latitude provided as a float (eg. -1.25). Latitude must be expressed as WGS84, i.e. [-90..+90]}
}
\value{
integer The corresponding PRIOGrid row id
}
\description{
The row (int) of the PRIOGrid structure corresponding to a latitude (float)
}
\examples{
pg_row(27.1)
pg_row(-24.5)
}
